requirements.txt
fire==0.1.2
kafka-python==1.3.5
six==1.11.0
